+++ packages/app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # app
 
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies [18fba21]
+- Updated dependencies [18fba21]
+  - @axis-backstage/plugin-jira-dashboard-common@1.9.0
+  - @axis-backstage/plugin-jira-dashboard@1.13.0
+
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies [39b1dbf]
+  - @axis-backstage/plugin-jira-dashboard-common@1.8.0
+  - @axis-backstage/plugin-jira-dashboard@1.12.0
+
 ## 0.10.0
 
 ### Minor Changes