+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,38 @@
 # backend
 
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies [eab8f06]
+  - @axis-backstage/plugin-jira-dashboard-backend@4.0.1
+
+## 0.7.7
+
+### Patch Changes
+
+- Updated dependencies [6c9c4b6]
+- Updated dependencies [18fba21]
+- Updated dependencies [18fba21]
+- Updated dependencies [18fba21]
+  - @axis-backstage/plugin-jira-dashboard-backend@4.0.0
+  - app@0.10.2
+
+## 0.7.6
+
+### Patch Changes
+
+- Updated dependencies [39b1dbf]
+  - @axis-backstage/plugin-jira-dashboard-backend@3.1.0
+  - app@0.10.1
+
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies [b6b406c]
+  - @axis-backstage/plugin-jira-dashboard-backend@3.0.0
+
 ## 0.7.4
 
 ### Patch Changes